feat(pet): 可靠模型切换、收藏菜单与测试 API
修复菜单/API 切换阻塞与 WebGL 污染，支持收藏角色换肤流程、空闲预加载与 HTTP 测试接口，完善 live2d-only 桌宠分支。

+++ src-tauri/src/db/sessions.rs
@@ -56,7 +56,7 @@ pub fn background_ms_for_date(db: &Connection, date: NaiveDate) -> Result<u64, r
         } else if let Ok(s) = chrono::DateTime::parse_from_rfc3339(&started) {
             let start = s.with_timezone(&Local);
             if start.date_naive() == date {
-                total += (Local::now() - start).num_milliseconds().max(0) as i64;
+                total += (Local::now() - start).num_milliseconds().max(0);
             }
         }
     }
