+++ src-tauri/src/db/usage.rs
@@ -101,7 +101,7 @@ fn ms_for_date(db: &Connection, table: &str, date: NaiveDate) -> Result<u64, rus
         } else if let Ok(s) = chrono::DateTime::parse_from_rfc3339(&started) {
             let start = s.with_timezone(&Local);
             if start.date_naive() == date {
-                total += (Local::now() - start).num_milliseconds().max(0) as i64;
+                total += (Local::now() - start).num_milliseconds().max(0);
             }
         }
     }
